Add divided renders in new files for better readable and organization
