+++ index.html
@@ -33,6 +33,8 @@
     </div>
     <main id="root"></main>
     <script type="module" src="dom-ul.js"></script>
-    <script defer type="module" src="app.js"></script>
+    <script type="module" src="app.js"></script>
+    <script type="module" src="detail.js"></script>
+    <script type="module" src="dashboard.js"></script>
   </body>
 </html>
