+++ style.css
@@ -1,118 +1,119 @@
 :root {
-    --color-dark-blue: hsl(209, 23%, 22%);
-    --color-very-dark-blue: hsl(207, 26%, 17%);
-    --color-mode-very-dark-blue:hsl(200, 15%, 8%);
-    --color-dark-grey: hsl(0, 0%, 52%);
-    --color-very-light-grey:hsl(0, 0%, 98%);
-    --color-white: hsl(0, 0%, 100%);
-
+  --color-dark-blue: hsl(209, 23%, 22%);
+  --color-very-dark-blue: hsl(207, 26%, 17%);
+  --color-mode-very-dark-blue: hsl(200, 15%, 8%);
+  --color-dark-grey: hsl(0, 0%, 52%);
+  --color-very-light-grey: hsl(0, 0%, 98%);
+  --color-white: hsl(0, 0%, 100%);
 }
 
 * {
-    box-sizing: border-box;
-    padding: 0;
+  box-sizing: border-box;
+  padding: 0;
 }
 
 body {
-    color: var(--color-mode-very-dark-blue);
-    font-family: 'Poppins', sans-serif;
-    
-    font-size: 14px;
+  color: var(--color-mode-very-dark-blue);
+  font-family: "Poppins", sans-serif;
+
+  font-size: 14px;
 }
 .filters,
 main {
-    
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 32px;
+  max-width: 1200px;
+  margin: 0 auto;
+  padding: 0 32px;
 }
 
 header {
-    background-color: var(--color-white);
-    margin-bottom: 32px;
+  background-color: var(--color-white);
+  margin-bottom: 32px;
 }
 
 .header-content {
-    display: flex;
-    justify-content: space-between;
-    margin: 0 auto;
-    padding: 32px 0;
-    max-width: 1200px;
+  display: flex;
+  justify-content: space-between;
+  margin: 0 auto;
+  padding: 32px 0;
+  max-width: 1200px;
 }
 
 header button {
-    background: none;
-    border: none;
-    cursor: pointer;
-    font-size: 24px;
+  background: none;
+  border: none;
+  cursor: pointer;
+  font-size: 24px;
 }
 
 .filters {
-    display: flex;
-    justify-content: space-between;
-    margin-bottom: 30px;
+  display: none;
+}
+
+.filters.active {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 30px;
 }
 
 #query {
-    background-color: var(--color-mode-very-dark-blue);
-    border: none;
-    box-shadow: 2px 3px 18px 5px rgba(66, 68, 90, 1);
-    padding: 15px 30px;
+  background-color: var(--color-mode-very-dark-blue);
+  border: none;
+  box-shadow: 2px 3px 18px 5px rgba(66, 68, 90, 1);
+  padding: 15px 30px;
 }
 
 main {
-    background-color: var(--color-white);
+  background-color: var(--color-white);
 }
 
 ul {
-    display: grid;
-    grid-template-columns: 1fr 1fr 1fr 1fr;
-    gap: 70px;
-    list-style: none;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  gap: 70px;
+  list-style: none;
 }
 
 li {
-    box-shadow: 2px 3px 18px 5px rgba(66, 68, 90, 1);
-    border-radius: 10px;
-    /* border-color: beige; */
-    overflow: hidden;
+  box-shadow: 2px 3px 18px 5px rgba(66, 68, 90, 1);
+  border-radius: 10px;
+  /* border-color: beige; */
+  overflow: hidden;
 }
 
 li a {
-    color: var(--color-white);
-    text-decoration: none;
+  color: var(--color-white);
+  text-decoration: none;
 }
 
 li img {
-    aspect-ratio: 160/98;
-    width: 100%;
-
+  aspect-ratio: 160/98;
+  width: 100%;
 }
 
 .info-container {
-    padding: 35px 20px;
+  padding: 35px 20px;
 }
 
 .country-name {
-    display: block;
-    font-size: 18px;
-    margin-bottom: 10px;
+  display: block;
+  font-size: 18px;
+  margin-bottom: 10px;
 }
 
 @media only screen and (max-width: 1000px) {
-    ul {
-        grid-template-columns: 1fr 1fr 1fr;
-    }
+  ul {
+    grid-template-columns: 1fr 1fr 1fr;
+  }
 }
 
 @media only screen and (max-width: 600px) {
-    ul {
-        grid-template-columns: 1fr 1fr;
-    }
+  ul {
+    grid-template-columns: 1fr 1fr;
+  }
 }
 
 @media only screen and (max-width: 400px) {
-    ul {
-        grid-template-columns: 1fr;
-    }
+  ul {
+    grid-template-columns: 1fr;
+  }
 }
